refactor(#1588): split retry tests to sibling file; tighten snappy oracle
Follow-ups from the #1588 review (epic #1135 campsite rule + a test nit).

1. Relocate the four #1588 retry/transient guards out of the inline
   `mod tests` in reader/block_io.rs into a sibling test file
   reader/block_io_retry_tests.rs, wired via
   `#[cfg(test)] #[path = "block_io_retry_tests.rs"] mod retry_tests;`
   (same pattern as scan_stream_windowed_tests.rs; `use super::*` resolves
   to block_io's private `retry_transient_once` / `is_transient_io`). Only
   the #1588-introduced tests moved; pre-existing tests are untouched. This
   drops block_io.rs 1983 -> 1846 lines. The residual over-threshold growth
   vs base (+19 lines) is pure #1588 production retry code (out of scope to
   change), and compression.rs remains over-and-grown from the #1588 snappy
   production+tests (also out of scope here), so the file-size ratchet still
   requires CQLITE_ALLOW_FILE_GROWTH=1 (acknowledged per epic #1135). No
   production logic changed.

2. Tighten test_snappy_decode_is_strict_raw_only_no_format_guessing: the
   ambiguous raw-vs-framed chunk must make strict raw decode ERROR (None),
   not merely differ from the framed guess. assert!(got.is_none()).
